After removing most of my authentication code, I can now create a user. I am just gonna have to build that stuff back in.
