+++ Pawfect-Harmony/src/main/java/org/launchcode/Pawfect/Harmony/models/dto/LoginFormDTO.java
@@ -21,16 +21,8 @@ public class LoginFormDTO {
         return username;
     }
 
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
     public String getPassword() {
         return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
     }
 
 }
