--- Pawfect-Harmony/src/main/java/org/launchcode/Pawfect/Harmony/models/dto/RegisterFormDTO.java
@@ -1,40 +1,0 @@
-package org.launchcode.Pawfect.Harmony.models.dto;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-public class RegisterFormDTO extends LoginFormDTO {
-
-
-    private String Email;
-
-    private String phone;
-
-    private String verifyPassword;
-
-    public String getEmail() {
-        return Email;
-    }
-
-    public void setEmail(String email) {
-        Email = email;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
-
-    public String getVerifyPassword() {
-        return verifyPassword;
-    }
-
-    public void setVerifyPassword(String verifyPassword) {
-        this.verifyPassword = verifyPassword;
-    }
-
-}
