+++ Pawfect-Harmony/src/main/resources/templates/index.html
@@ -7,11 +7,9 @@
     <div>
         <h1>Welcome to Pawfect Harmony</h1>
         <h2>Your pet adoption Partner</h2>
-        <h3>Are you looking to:</h3>
     </div>
     <div>
-        <a href="/search" class="btn btn-primary btn-lg active" role="button" aria-pressed="true">Adopt a Pet</a>
-        <a href="/login" class="btn btn-primary btn-lg active" role="button" aria-pressed="true">Put a pet up for adoption</a>
+        <a href="/login" class="btn btn-primary btn-lg active" role="button" aria-pressed="true">Please Login to begin</a>
     </div>
