+++ Pawfect-Harmony/src/main/resources/templates/login.html
@@ -20,7 +20,7 @@
   <input type="submit" class="btn btn-primary" value="Log In" />
 </form>
 
-<p>Don't have an account? <a href="/register">Register for one.</a></p>
+<p>Don't have an account? <a href="user/create">Register for one.</a></p>
 
 </body>
 </html>
