+++ Pawfect-Harmony/src/main/resources/templates/user/create.html
@@ -5,37 +5,37 @@
 
 <form method="post" style="max-width:600px;">
     <div class="form-group">
-        <label th:for="name">User Name</label>
+        <label>User Name</label>
         <input class="form-control" th:field="${user.username}" />
         <span th:errors="${user.username}" class="error"></span>
     </div>
 
     <div class="form-group">
-        <label th:for="name">Password</label>
-        <input class="form-control" th:field="${user.password}" />
+        <label>Password</label>
+        <input type="password" class="form-control" th:field="${user.password}" />
         <span th:errors="${user.password}" class="error"></span>
     </div>
 
     <div class="form-group">
-        <label th:for="location">First Name</label>
+        <label>First Name</label>
         <input class="form-control" th:field="${user.firstName}" />
         <span th:errors="${user.firstName}" class="error"></span>
     </div>
 
     <div class="form-group">
-        <label th:for="location">Last Name</label>
+        <label>Last Name</label>
         <input class="form-control" th:field="${user.lastName}" />
         <span th:errors="${user.lastName}" class="error"></span>
     </div>
 
     <div class="form-group">
-        <label th:for="location">Email</label>
+        <label>Email</label>
         <input class="form-control" th:field="${user.email}" />
         <span th:errors="${user.email}" class="error"></span>
     </div>
 
     <div class="form-group">
-        <label th:for="location">Phone</label>
+        <label>Phone</label>
         <input class="form-control" th:field="${user.phone}" />
         <span th:errors="${user.phone}" class="error"></span>
     </div>
